+++ src/sycl-acc/jacobi.cpp
@@ -7,7 +7,7 @@ using namespace cl::sycl;
 // Initialises the Jacobi solver
 void jacobi_init(const int x,             //
                  const int y,             //
-                 const int halo_depth,    //
+                 const size_t halo_depth, //
                  const int coefficient,   //
                  const double rx,         //
                  const double ry,         //
@@ -28,11 +28,11 @@ void jacobi_init(const int x,             //
     h.parallel_for<class jacobi_init>(range<1>(x * y), [=](id<1> idx) {
       const auto kk = idx[0] % x;
       const auto jj = idx[0] / x;
-      if (kk > 0 && kk < x - 1 && jj > 0 && jj < y - 1) {
+      if (kk > 0 && kk < size_t(x - 1) && jj > 0 && jj < size_t(y - 1)) {
         u0[idx[0]] = energy[idx[0]] * density[idx[0]];
         u[idx[0]] = u0[idx[0]];
       }
-      if (jj >= halo_depth && jj < y - 1 && kk >= halo_depth && kk < x - 1) {
+      if (jj >= halo_depth && jj < size_t(y - 1) && kk >= halo_depth && kk < size_t(x - 1)) {
         double densityCentre = (coefficient == CONDUCTIVITY) ? density[idx[0]] : 1.0 / density[idx[0]];
         double densityLeft = (coefficient == CONDUCTIVITY) ? density[idx[0] - 1] : 1.0 / density[idx[0] - 1];
         double densityDown = (coefficient == CONDUCTIVITY) ? density[idx[0] - x] : 1.0 / density[idx[0] - x];
@@ -48,15 +48,15 @@ void jacobi_init(const int x,             //
 }
 
 // Main Jacobi solver method.
-void jacobi_iterate(const int x,          //
-                    const int y,          //
-                    const int halo_depth, //
-                    SyclBuffer &uBuff,    //
-                    SyclBuffer &u0Buff,   //
-                    SyclBuffer &rBuff,    //
-                    SyclBuffer &kxBuff,   //
-                    SyclBuffer &kyBuff,   //
-                    double *error,        //
+void jacobi_iterate(const int x,             //
+                    const int y,             //
+                    const size_t halo_depth, //
+                    SyclBuffer &uBuff,       //
+                    SyclBuffer &u0Buff,      //
+                    SyclBuffer &rBuff,       //
+                    SyclBuffer &kxBuff,      //
+                    SyclBuffer &kyBuff,      //
+                    double *error,           //
                     queue &device_queue) {
   buffer<double, 1> error_temp{range<1>{1}};
   device_queue.submit([&](handler &h) {
